Made fixes in code. Announcements work now, however the UI for no announcements is still pending.
